Backend/requirements.txt
fastapi>=0.70.0
uvicorn[standard]>=0.18.0
httpx>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=0.8.11
google-cloud-storage>=2.10.0
